Fixed exception handling in connector file
Most except blocks were accessing the 'message' property of an Exception object. However, this property does not exist. This was never caught because a bug in the xforce module prevented HTTP status exceptions to be raised.
Instead, we now catch XForceError exceptions and we pass them directly to the BaseConnector.set_status_save_progress() method.